release/requirements.txt
pyclipper==1.1.0.post1
torchvision==0.4.1
numpy==1.17.3
tqdm==4.38.0
opencv_python==4.1.1.26
torch==1.3.0
configobj==5.0.6
h5py==2.10.0
Pillow==6.2.1
Polygon==3.0.5
PyQt5==5.13.2
scipy==1.3.2
